graphrag_mcp/requirements.txt
lightrag-hku>=1.4.0,<2.0.0
fastembed>=0.4.0,<1.0.0
mcp>=1.0.0
httpx>=0.27.0,<1.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
